Fix preround for Fraction and float types (#26)
Since #25, the definition of `preround` didn't match the description,
for `float` and `Fraction`: we want the incoming exponent to be the
target exponent. The preround machinery is free to return anything with
exponent equal to or lower than the given exponent.

This was still working because we were still (incorrectly) decrementing
the exponent passed to `preround` in the first place, passing one less
than the target exponent instead of the target exponent itself.

This PR fixes both issues.

Not a user-facing bug, since the functions exposed in the API continued
to work as expected.