Support exact formatting (#29)
[Pushing an ancient branch to GitHub, for self-review purposes]

This PR adds support for formatting in the case where both
`format.maximum_figures` and `format.minimum_exponent` are `None`. In
that case, formatting should either give a representation of the exact
value, or it should fail.

## Detailed changes 

- support the formatting case where both `format.maximum_figures` and
`format.minimum_exponent` are `None`: in that case, formatting should
either give a representation of the exact value, or fail
- support an exponent of `None` in
`IntermediateForm.from_signed_fraction`
- make exponent optional in `preround`: if `exponent` is `None`, then
either an exact conversion to the intermediate form is performed, or if
that isn't possible, `ValueError` is raised.