Simplify handling of the formatted sign for zeros (#75)
This PR introduces the following changes to the `FormatSpecification`
class:

- add a flag `signed_zero` indicating whether the target format permits
negative zeros
- remove `negative_zero_sign` and `positive_zero_sign` attributes

The `format` implementation is updated to use a new support method
`force_unsigned_zero` on the `IntermediateForm` class.

There's a conceptual shift here: we think of the `z` flag as targeting
the _rounding_ part of the formatting operation rather than the
presentation part.