Implement formatting for zeros (#79)
This PR implements formatting for zeros. The hard part of this is
deciding what exponent to use.

For fixed-precision formatting, the choice is clear: we use the same
target exponent when formatting both nonzero and zero numbers.

For formatting a zero to a given number of significant digits, we choose
the exponent so that the formatted output has the expected number of
significant digits and so that in scientific notation the exponent would
be zero. So for example when formatting to 3 significant digits, any
zero ends up as `0.00`, or equivalently `0.00e+0`.

If targeting a format that has both a limit on the number of significant
digits _and_ on the precision, we take the max of the implied exponents.
So for example if `places = 3` and `figures = 5`, the result is `0.000`
rather than `0.0000`. This works well for the particular case `places =
0`, where all target values are integral.