Preserve natural exponent in preround (#87)
Closes #84 

This PR updates `preround` for `float` and `Fraction` to ensure that it
uses the "natural" exponent, when possible, for the resulting
`IntermediateForm` object.

Here we define the "natural exponent" of a float or fraction `x` to be
the largest nonpositive integer `e` such that `x` is an integer multiple
of `10**e`. The same applies to `int`, where the natural exponent of any
integer is `0`.

We also add tests for `preround` for `float`, `Fraction` and `int`.