Refactor preround implementations (#95)
This PR refactors some of the prerounding support:

- Remove `IntermediateForm.from_signed_fraction`: the name was
misleadingly general (it didn't indicate that there was prerounding
happening), and with this refactor the code is only used by the
`preround` implementation for `fractions.Fraction` (previously it was
used for `float`, too)
- Simplify the `preround` implementation for `float`, making it more
efficient.
- Add `reciprocal_as_decimal` function, which is the main algorithm used
for the `fractions.Fraction` preround implementation.
- Rename `log5exact` to `reciprocal_as_decimal`, since that's what we
really care about being able to compute. The power-of-five detection is
just an implementation detail.