Refactor and rename round_for_format (#98)
This PR:

- Renames `round_for_format` to `round_to_format`, for consistency with
other `round_to` functions.
- Updates the signature of `round_to_format` for consistency with other
`round_to` functions: the first argument is now called `number`, and the
rounding mode has a default of ties-to-even.
- Simplifies `round_for_format` to remove special-casing of zeros: zeros
now follow the usual rounding logic, and the presentation stage of the
formatting is adjusted to enforce a minimum number of zeros after the
point.